requirements.txt
flask
flasgger
selenium
beautifulsoup4
python-dotenv
